Create/update version tag when making tarballs
This will warn if the version of ascron is already tagged.

If git isn't present, or not in a working directory, no
attempt to tag will be made.  This is important so that
make install works from unpacked tarballs.

Use ascron version, not Perl version for man page footer.

Support user-specified signing key(s) for tarballs.

Include INSTALL in tarball.